Joint density function for two conditionally Weibull random variables, marginalized over the frailty term.

If correlated, the function computes the marginal joint density of Y and Z by marginalizing over the support of K. For computational efficiency, it integrates from epsilon to int.upper, where epsilon can be 0 and int.upper can be Inf. Small deviations from 0 and Inf will approximate the true integral.

f.YZ.Weibull(y, z, parameters_x, parameters_y, gamma_shape, gamma_rate,
  epsilon, control_list, int.upper, correlated = TRUE)

Arguments

y
A number at which to evaluate the density \(Y = y\)

z
A number at which to evaluate the density \(X = z-y\)

parameters_x
A list, with an element named lambda (\(\lambda_x\)), and an element named p (\(p_x\)).

parameters_y
A list, with an element lambda (\(\lambda_y\)), and an element named p (\(p_y\)).

gamma_shape
A number, the shape parameter of Gamma frailty distribution

gamma_rate
A number, the rate parameter of Gamma frailty distribution

epsilon
A float, the lower bound for support of K

control_list
A list, with an element named rel.tol to be passed to `distrExIntegrate` as the rel.tol argument (see ?distrExIntegrate)

int.upper
A float, the upper bound for support of K

correlated
If FALSE, then \(X\) and \(Y\) are independent. If TRUE, then there is a shared frailty term \(K\) that induces correlation between \(X\) and \(Y\).

Value

joint density function evaluated at Y=y, and X=z-y.

Details

\(X \sim Weibull(shape=p_{x}, rate=\lambda_{x} \cdot k^{p_{x}})\)
\(Y \sim Weibull(shape=p_{y}, rate=\lambda_{y} \cdot k^{p_{y}})\)
\(Z := X + Y \to X = Z - Y\)
\(K \sim Gamma(shape=gamma_shape, rate=gamma_rate)\)